backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
cloudinary
psycopg2-binary